+++ genmonlib/generac_HPanel.py
@@ -1966,7 +1966,7 @@ class HPanel(GeneratorController):
 
             if not IsFile and self.RegisterIsBaseRegister(Register, Value):
                 # TODO validate register length
-                self.Registers[Register] = Value
+                self.Holding[Register] = Value
             elif not IsFile and self.RegisterIsStringRegister(Register):
                 # TODO validate register string length
                 self.Strings[Register] = Value
@@ -3321,11 +3321,11 @@ class HPanel(GeneratorController):
 
             RegList = []
 
-            Regs["Num Regs"] = "%d" % len(self.Registers)
+            Regs["Num Regs"] = "%d" % len(self.Holding)
 
-            Regs["Holding Registers"] = RegList
+            Regs["Holding"] = RegList
             # display all the registers
-            temp_regsiters = self.Registers
+            temp_regsiters = self.Holding
             for Register, Value in temp_regsiters.items():
                 RegList.append({Register: Value})